# Conflicts:
#	mgs/pom.xml

#OtherChanges
- Award related changes
- Other bug fixing

+++ mgs/src/main/angular/src/app/modules/student/student-activity/saved-actitvity/saved-actitvity.component.ts
@@ -29,21 +29,27 @@ export class SavedActitvityComponent implements OnInit {
   activityDate: any;
   file = [];
   url = '';
-  activityType = "All";
+  activityType = 'Submitted';
+  loader:boolean = false;
 
   constructor(private formBuilder: FormBuilder, private studentService: StudentService, private alertService: AlertService) { }
 
   ngOnInit() {
     this.studentInfo = JSON.parse(localStorage.getItem('user_info'));
-    this.studentId = this.studentInfo['student'].id
+    // tslint:disable-next-line:no-string-literal
+    this.studentId = this.studentInfo['student'].id;
+    // tslint:disable-next-line:no-string-literal
     this.schoolId = this.studentInfo['student'].schoolId;
 
     // this.getStudentSavedActivities(this.studentInfo['student'].id)
-    this.getStudentSavedActivities(this.studentId);
-    this.getStudentSubmittedActivities(this.studentId);
-    this.getStudentAllActivities(this.studentId);
+
+    // this.getStudentSavedActivities(this.studentId);
+    // this.getStudentSubmittedActivities(this.studentId);
+    // this.getStudentAllActivities(this.studentId);
+    // this.getStudentReviewedActivities(this.studentId);
 
     // this.getStudentActivity(this.schoolId);
+    this.activityView(this.activityType);
 
     this.savedActivityForm = this.formBuilder.group({
       savedActivityName: [''],
@@ -53,13 +59,15 @@ export class SavedActitvityComponent implements OnInit {
       coachId: [''],
       attachment: [''],
       id: ['']
-    })
+    });
   }
 
   // to get the list of SAVED Activities of student
   getStudentSavedActivities(studentId) {
     this.studentService.getSavedActivity(studentId).subscribe((res) => {
       this.savedActivitiesArr = res;
+      console.log(res);
+      this.loader = false;
     },
       (err) => console.log(err));
   }
@@ -68,6 +76,7 @@ export class SavedActitvityComponent implements OnInit {
   getStudentSubmittedActivities(studentId) {
     this.studentService.getSubmittedActivity(studentId).subscribe((res) => {
       this.submittedActivitiesArr = res;
+      this.loader = false;
     },
       (err) => console.log(err)
     );
@@ -76,29 +85,45 @@ export class SavedActitvityComponent implements OnInit {
   // to get the list of ALL Activities of student
   getStudentAllActivities(studentId) {
     this.studentService.getAllActivity(studentId).subscribe((res) => {
+      console.log(res);
       this.allActivitiesArr = res;
+      this.allActivitiesArr = this.allActivitiesArr.filter((e) => (e.activityStatus != "SavedByTeacher"));
+      this.loader = false;
     },
       (err) => console.log(err)
     );
   }
 
+  // to get the list of REVIEWED Activities of student
+  getStudentReviewedActivities(studentId) {
+    this.studentService.getReviewedActivity(studentId).subscribe((res) => {
+      this.reviewedActivitiesArr = res;
+      this.loader = false;
+    },
+      (err) => console.log(err));
+
+  }
+
   // on click of edit button
-  editSavedActivity(activity) {
+  editSavedActivity(e, activity) {
     // this.activityName = activity.activityName;
     // this.activityTeacher = activity.teacherName;
+    e.stopPropagation();
+    console.log(activity);
+    this.activityId = activity.activityId;
 
     const date = this.getDate(activity.dateOfActivity);
     if ((date.getMonth() + 1) < 10) {
       if (date.getDate() < 10) {
-        this.savedActivityForm.controls['dateOfActivity'].patchValue(date.getFullYear() + '-' + ('0' + (date.getMonth() + 1)) + '-' + ('0' + date.getDate()));
+        this.savedActivityForm.controls.dateOfActivity.patchValue(date.getFullYear() + '-' + ('0' + (date.getMonth() + 1)) + '-' + ('0' + date.getDate()));
       } else {
-        this.savedActivityForm.controls['dateOfActivity'].patchValue(date.getFullYear() + '-' + ('0' + (date.getMonth() + 1)) + '-' + date.getDate());
+        this.savedActivityForm.controls.dateOfActivity.patchValue(date.getFullYear() + '-' + ('0' + (date.getMonth() + 1)) + '-' + date.getDate());
       }
     } else {
       if (date.getDate() < 10) {
-        this.savedActivityForm.controls['dateOfActivity'].patchValue(date.getFullYear() + '-' + (date.getMonth() + 1) + '-' + ('0' + date.getDate()));
+        this.savedActivityForm.controls.dateOfActivity.patchValue(date.getFullYear() + '-' + (date.getMonth() + 1) + '-' + ('0' + date.getDate()));
       } else {
-        this.savedActivityForm.controls['dateOfActivity'].patchValue(date.getFullYear() + '-' + (date.getMonth() + 1) + '-' + date.getDate());
+        this.savedActivityForm.controls.dateOfActivity.patchValue(date.getFullYear() + '-' + (date.getMonth() + 1) + '-' + date.getDate());
       }
     }
 
@@ -106,33 +131,46 @@ export class SavedActitvityComponent implements OnInit {
       activityId: activity.activityId,
       description: activity.description,
       id: activity.id,
-    })
+    });
+
     console.log(this.savedActivityForm.value);
+
 
     this.getStudentActivity(this.schoolId);
     return this.editActivityShow = true;
   }
 
   // to SUBMIT the activity
-  submitSavedActivity(index) {
-    const activityId = this.savedActivitiesArr[index].id;
-    this.studentService.submitActivity(activityId).subscribe((res) => {
-      console.log(res);
-      this.alertService.showSuccessToast('Activity Submitted !');
-      // splice the selected activity from savedActivity array and push it into submit activity array
-    },
-      (err) => console.log(err)
-    );
+  submitSavedActivity(e, index) {
+    const submit = confirm("Do you want to submit ?");
+    if(submit){
+      e.stopPropagation();
+      const activityId = this.savedActivitiesArr[index].id;
+  
+      console.log(this.submittedActivitiesArr);
+      this.studentService.submitActivity(activityId).subscribe((res) => {
+        console.log(res);
+        this.submittedActivitiesArr = [...this.submittedActivitiesArr, ...this.savedActivitiesArr.splice(index, 1)];
+        this.alertService.showSuccessToast('Activity Submitted !');
+      },
+        (err) => console.log(err)
+      );      
+    }
   }
 
-  deleteSavedActivity(activity) {
-    const activityId = activity.id;
-    console.log(activityId);
-    this.studentService.deleteActivity(activityId).subscribe((res) => {
-      console.log(res);
-      this.alertService.showSuccessToast('Activity Deleted !');
-    },
-    (err) => console.log(err));
+  deleteSavedActivity(e, activity, i) {
+    const submit = confirm("Do you want to delete ?");
+    if(submit){
+      e.stopPropagation();
+      const activityId = activity.id;
+      console.log(activityId);
+      this.studentService.deleteActivity(activityId).subscribe((res) => {
+        console.log(res);
+        this.savedActivitiesArr.splice(i, 1);
+        this.alertService.showSuccessToast('Activity Deleted !');
+      },
+        (err) => console.log(err));
+    }
   }
 
   onCancel() {
@@ -161,7 +199,7 @@ export class SavedActitvityComponent implements OnInit {
   onFileSelect(event) {
     if (event.target.files.length > 0) {
       this.file = [...event.target.files];
-      this.savedActivityForm.value['attachment'] = this.file;
+      this.savedActivityForm.value.attachment = this.file;
       console.log(this.file);
     }
   }
@@ -169,15 +207,15 @@ export class SavedActitvityComponent implements OnInit {
   // to UPDATE the saved activity
   updateActivity() {
     const formData = new FormData();
-    let date = new Date(this.savedActivityForm.value.dateOfActivity);
-    let activityDate = date.getFullYear() + "/" + (date.getMonth() + 1) + "/" + date.getDate();
+    const date = new Date(this.savedActivityForm.value.dateOfActivity);
+    const activityDate = date.getFullYear() + '/' + (date.getMonth() + 1) + '/' + date.getDate();
 
     formData.append('studentId', this.studentInfo['student'].id);
     formData.append('activityId', this.savedActivityForm.value.activityId);
     formData.append('coachId', this.savedActivityForm.value.coachId);
     formData.append('dateOfActivity', activityDate);
     formData.append('description', this.savedActivityForm.value.description);
-    formData.append('id', this.savedActivityForm.value.id)
+    formData.append('id', this.savedActivityForm.value.id);
 
     if (this.savedActivityForm.value.attachment.length > 0) {
       this.savedActivityForm.value.attachment.forEach((element, index) => {
@@ -185,7 +223,7 @@ export class SavedActitvityComponent implements OnInit {
       });
     }
 
-    this.studentService.addActivity("/api/students/activities", formData).subscribe(
+    this.studentService.addActivity('/api/student/activities', formData).subscribe(
       (res) => {
         console.log(res);
         this.alertService.showSuccessToast('Activity Updated !');
@@ -196,10 +234,24 @@ export class SavedActitvityComponent implements OnInit {
 
   activityView(event) {
     this.activityType = event;
+    this.loader = true;
+    switch (this.activityType) {
+      case "All": this.getStudentAllActivities(this.studentId);  
+        break;
+
+      case "Saved" : this.getStudentSavedActivities(this.studentId);
+      break;
+
+      case "Reviewed" : this.getStudentReviewedActivities(this.studentId);
+      break;
+      
+      case "Submitted" : this.getStudentSubmittedActivities(this.studentId);
+      break;
+    }
   }
 
   getDate(date) {
-    return new Date(date)
+    return new Date(date);
   }
 
 }
